Refactor search UI into feature components.
Extracts search header/suggestions, results list, and discover sections into reusable components to reduce screen complexity.
